refactor(test): Update biomech tests.
Update some error codes to reflect their error better and disabled file delete when testing. Tests still generate images but that will be solved later.

Issue: https://github.sfu.ca/bfraser/415-HHA-Haiti/issues/714

+++ server/src/utils/unlinkImage.ts
@@ -3,9 +3,16 @@ const fs = require('fs');
 import { logger } from '../logger';
 
 export const deleteUploadedImage = (filepath: string) => {
-  fs.unlink(filepath, (error: any) => {
-    error
-      ? logger.error(error.message)
-      : logger.debug('DELETED IMAGE WITH FILEPATH: '.concat(filepath));
-  });
+  //TODO: find better way to deal with this
+  if (process.env.NODE_ENV !== 'test') {
+    fs.unlink(filepath, (error: any) => {
+      error
+        ? logger.error(error.message)
+        : logger.debug('DELETED IMAGE WITH FILEPATH: '.concat(filepath));
+    });
+  } else {
+    logger.debug(
+      'In not in test environment, would have DELETED IMAGE WITH FILEPATH: '.concat(filepath),
+    );
+  }
 };
